improve(entrypoint): spool mode setup improvements
Handle DNS01_SPOOL and SPOOL_DIR definitions correctly

SPOOL_DIR if set, tells the spooler where the request/response files
will live - if unset, it defaults to a /spool subdirectory at the
spool.sh instance's path.

DNS01_SPOOL is handled at entry point and is the path where spool.sh
should be copied to and called from. If unset, it will default first to
SPOOL_DIR if that is set, otherwise it will default to DNS01_PATH.

The spooler doesn't attempt to create the spool directory and aborts if
it's not present. The entry point does attempt to create the directory,
first checking against SPOOL_DIR and then against DNS01_SPOOL/spool.

In this way the two variables can be combined or left to defaults to
control where the script and the spool directory are expected to live
in spooled mode - spool.sh stays decoupled from the image/mode logic.